% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tokenizeData}
\alias{tokenizeData}
\title{Tokenize data}
\usage{
tokenizeData(d)
}
\arguments{
\item{d}{a setence}
}
\value{
a list with words
}
\description{
Remove al special characters and break it by word
}

